02_dataform_bqml/bootstrap/requirements.txt
google-cloud-dataform
google-api-core